refactoring checksum logic
following es6
