+++ functions/checksum/checksum.js
