+++ functions/index.js
@@ -6,8 +6,8 @@ const checksum = require('./modules/checksum');
 
 admin.initializeApp(functions.config().firebase);
 
-exports.test = test;
+exports.test = functions.https.onRequest((req,res) => test.test(req, res));
 
-exports.response = response;
+exports.response = functions.https.onRequest((req,res) => response.response(req, res));;
 
-exports.checksum = checksum;
+exports.checksum = functions.https.onRequest((req,res) => checksum.checksum(req, res));;
