+++ functions/modules/test.js
@@ -1,8 +1,6 @@
-const functions = require('firebase-functions');
-
-export default functions.https.onRequest((req, res) => {
+exports.test = (req, res) => {
     //Test https function
     res.writeHead(200);
     res.write("OK");
     res.end();
-});
+};
